feat: added functionality that allows user to create a new note.
Everything is setup to create a new note. All that is left is having the front-end send a request to the backend with the proper data to save to the database.

+++ src/main/java/com/raistmere/notetakingwebapp/model/NoteModel.java
@@ -4,13 +4,13 @@ public class NoteModel {
 
     private int id;
 
-    private int userId;
+    private Long userId;
     private String title;
     private String note;
 
     public NoteModel() {}
 
-    public NoteModel(int id, int userId, String note) {
+    public NoteModel(int id, Long userId, String note) {
         this.id = id;
         this.userId = userId;
         this.note = note;
@@ -24,11 +24,11 @@ public class NoteModel {
         this.id = id;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
